+++ packages/components/grid/src/filter-column.ts
@@ -1,5 +1,6 @@
 import type { PropertyValues, TemplateResult } from 'lit';
 import { getNameByPath, getValueByPath } from '@sl-design-system/shared';
+import { localized, msg } from '@lit/localize';
 import { html } from 'lit';
 import { property, state } from 'lit/decorators.js';
 import { GridColumn } from './column.js';
@@ -12,6 +13,7 @@ export interface GridFilterOption {
   value?: unknown;
 }
 
+@localized()
 export class GridFilterColumn extends GridColumn {
   /** The internal options if none are provided. */
   @state() internalOptions?: GridFilterOption[];
@@ -56,10 +58,16 @@ export class GridFilterColumn extends GridColumn {
       // No options were provided, so we'll create a list of options based on the column's values
       this.internalOptions = this.grid?.items
         ?.reduce((acc, item) => {
-          const value = getValueByPath(item, this.path);
+          let value = getValueByPath(item, this.path),
+            label = value?.toString() ?? '';
+
+          if (value === null || value === undefined || (typeof value === 'string' && value.trim() === '')) {
+            label = msg('Blank');
+            value = '';
+          }
 
           if (value !== null && !acc.some(option => option.value === value)) {
-            acc.push({ label: value?.toString() ?? '', value });
+            acc.push({ label, value });
           }
 
           return acc;